requirements.txt
asgiref==3.6.0
Django==4.1.5
django-ninja==0.20.0
pydantic==1.10.4
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
